Tighten up HTTP errors
